Delete From Data Base Fixed and Refund Added

+++ src/main/java/com/github/daymoon/Cart.java
@@ -65,15 +65,16 @@ public class Cart {
     public void DeleteFromDataBase(){
 
         try(Connection conn = DriverManager.getConnection("jdbc:sqlite:C:\\Java\\mini-ecommerce\\DataBase\\mini-ecommerce-database.db")){
-            String sql = "DELETE FROM Cart WHERE productId = ?";
+            String sql = "DELETE FROM Cart WHERE productId = ? AND userId = ?";
             PreparedStatement pst = conn.prepareStatement(sql);
             pst.setInt(1, this.productId);
+            pst.setInt(2, this.buyerId);
 
             int deleted = pst.executeUpdate();
 
             if(deleted>0){
             }else{
-                System.out.println("Failed to Find the Product");
+                System.out.println("Failed to Find the Cart Product");
             }
 
         }catch(Exception e){
